requirements.txt
beautifulsoup4>=4.12.2
curl_cffi
DateTime>=4.9
icalendar>=4.0.9
icalevents>=0.1.29
python-dateutil>=2.8.2
pytz>=2021.3
PyYAML>=6.0.1
requests>=2.31.0
urllib3>=2.0.7
jinja2>=3.1.2
lxml>=4.9.4
pycryptodome>=3.20.0
typing_extensions>=4.12.2
pypdf>=5.1.0
pdfminer.six>=20221105
homeassistant